Log to file and console; hacky replace print with log
Everything that was a print call now calls plog instead, which routes
the message through the root logger. Due to the new logger configuration
the end result is that the message is printed to the console and saved
in the minerva.log file.

We need to distinguish between info and error plog calls, probably by
calling into the logger directly.